Add shake to roll option
The device motion API is deprecated since it can be used for tracking,
but this adds support where it's available.

The code isn't perfect -- it'd be better to determine the vector of
movement and see that that gets reversed a few times rather than looking
at each cardinal direction and rotation axis individually -- but it does
the job.

Maybe in future this will be useful again, if browsers implement an
opt-in mechanism for device motion per website, like for location
details etc.